Fix for bom ftp sync issue

+++ sss/management/commands/sync_ftp_bom.py
@@ -129,21 +129,22 @@ class Command(BaseCommand):
                                 logger.error(traceback.format_exc())
                                 pass
                             continue
-                        
+                                                
                         logger.info("Copying File "+temp_file_path)
                         # Copy GZ File
                         shutil.copyfile(temp_file_path, local_file_path)
                         # Copy NC File
 
                         # This check is important to confirm the file copy to shared storage and is not broken
-                        shutil.copyfile(unzipped_file_path, os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path+".tmp.nc")))                        
+                        shutil.copyfile(unzipped_file_path, os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path+".tmp.nc")))            
                         file = gdal.Open(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path+".tmp.nc")))
                         file.GetGeoTransform()
-                        os.remove(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path)))
+                        if os.path.exists(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path))):
+                            os.remove(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path)))
                         os.rename(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path+".tmp.nc")), os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(unzipped_file_path)))
 
                         os.remove(temp_file_path)
-                        os.remove(unzipped_file_path)
+                        os.remove(unzipped_file_path)                    
                         
                     except Exception as e:
                         logger.error(f"Unzipping failed for {temp_file_name}")
@@ -167,15 +168,15 @@ class Command(BaseCommand):
                                 logger.error(traceback.format_exc())
                                 pass
                             continue
-
+                        
                         logger.info("Copying File "+temp_file_path)
-                        # This check is important to confirm the file copy to shared storage and is not broken
-                        shutil.copyfile(temp_file_path, local_file_path+".tmp.nc")                        
+                        # This check is important to confirm the file copy to shared storage and is not broken                                                                      
+                        shutil.copyfile(temp_file_path, local_file_path+".tmp.nc")
                         file = gdal.Open(local_file_path+".tmp.nc")
                         file.GetGeoTransform()
-
-                        os.remove(local_file_path)
-                        os.rename(local_file_path+".tmp.nc", local_file_path)
+                        shutil.copyfile(temp_file_path, os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(temp_file_path+".tmp.nc")))
+                        os.rename(os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(temp_file_path+".tmp.nc")), os.path.join(BOM_HOME_LOCAL, bom_ftp_directory, os.path.basename(temp_file_path)))
+                        os.remove(temp_file_path)                        
 
                     except Exception:
                         logger.error(f"File copy/delete failed for {temp_file_name}")
